Move observe service into it's own module. More cleanup.
