+++ e2e-tests/src/lib.rs
@@ -1,3 +1,1 @@
-#![deny(elided_lifetimes_in_paths)]
-
 mod observe;
